requirements.txt
streamlit==1.48.1
pandas==2.3.1
numpy==2.3.2
matplotlib==3.9.2
scipy==1.16.1
pyedflib==0.1.34
pdfplumber==0.11.7
python-docx==1.2.0
fpdf==1.7.2
Pillow==11.3.0
easyocr==1.7.2
opencv-python-headless==4.11.0.86
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
